[ADD] Models and endpoints are added
