+++ backend/src/routes/index.js
@@ -1,9 +1,12 @@
 const express = require('express');
 const router = express.Router();
-const { getProducts } = require('../controllers/productController');
+const productController = require('../controllers/productController');
+const cartController = require('../controllers/cartController');
 
-router.get('/products', getProducts);
-router.get('/product/:id', getProducts);
-router.post('/cart', getProducts);
+router.get('/products', productController.getProducts);
+
+router.get('/product/:id', productController.getProductById);
+
+router.post('/cart', cartController.addToCart);
 
 module.exports = router;
